+++ src/Service/LeantimeApiService.php
@@ -4,13 +4,14 @@ namespace App\Service;
 
 use App\Enum\IssueStatusEnum;
 use App\Exception\ApiServiceException;
-use App\Exception\EconomicsException;
 use App\Interface\DataProviderServiceInterface;
 use App\Model\Invoices\IssueData;
 use App\Model\Invoices\PagedResult;
 use App\Model\Invoices\ProjectData;
 use App\Model\Invoices\ProjectDataCollection;
 use App\Model\Invoices\VersionData;
+use App\Model\Invoices\VersionModel;
+use App\Model\Invoices\Versions;
 use App\Model\Invoices\WorklogData;
 use App\Model\Invoices\WorklogDataCollection;
 use App\Model\Planning\Assignee;
@@ -20,15 +21,6 @@ use App\Model\Planning\PlanningData;
 use App\Model\Planning\Project;
 use App\Model\Planning\SprintSum;
 use App\Model\Planning\Weeks;
-use App\Model\SprintReport\SprintReportData;
-use App\Model\SprintReport\SprintReportEpic;
-use App\Model\SprintReport\SprintReportIssue;
-use App\Model\SprintReport\SprintReportProject;
-use App\Model\SprintReport\SprintReportProjects;
-use App\Model\SprintReport\SprintReportSprint;
-use App\Model\SprintReport\SprintReportVersion;
-use App\Model\SprintReport\SprintReportVersions;
-use App\Model\SprintReport\SprintStateEnum;
 use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Component\Uid\Ulid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
@@ -36,11 +28,6 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 class LeantimeApiService implements DataProviderServiceInterface
 {
     private const API_PATH_JSONRPC = '/api/jsonrpc/';
-    private const NO_SPRINT = 'NoSprint';
-    private const PAST = 'PAST';
-    private const PRESENT = 'PRESENT';
-    private const FUTURE = 'FUTURE';
-
     private const LEANTIME_TIMEZONE = 'UTC';
 
     private static ?\DateTimeZone $leantimeTimeZone = null;
@@ -58,7 +45,6 @@ class LeantimeApiService implements DataProviderServiceInterface
         protected readonly float $weekGoalLow,
         protected readonly float $weekGoalHigh,
         protected readonly string $sprintNameRegex,
-        private readonly DateTimeHelper $dateTimeHelper,
     ) {
     }
 
@@ -72,7 +58,7 @@ class LeantimeApiService implements DataProviderServiceInterface
     /**
      * Get all projects, including archived.
      *
-     * @throws ApiServiceException|EconomicsException
+     * @throws ApiServiceException
      */
     public function getAllProjects(): mixed
     {
@@ -85,6 +71,8 @@ class LeantimeApiService implements DataProviderServiceInterface
      * @param string $projectId the ID of the project
      *
      * @return \stdClass the ticket status settings of the project
+     *
+     * @throws ApiServiceException
      */
     private function getProjectTicketStatusSettings(string $projectId): \stdClass
     {
@@ -97,9 +85,6 @@ class LeantimeApiService implements DataProviderServiceInterface
      * @param $projectId
      *
      * @return mixed
-     *
-     * @throws ApiServiceException
-     * @throws EconomicsException
      */
     public function getProjectWorklogs($projectId): mixed
     {
@@ -109,51 +94,7 @@ class LeantimeApiService implements DataProviderServiceInterface
     }
 
     /**
-     * Get all projects, including archived.
-     *
-     * @return SprintReportProjects array of SprintReportProjects
-     *
-     * @throws ApiServiceException|EconomicsException
-     */
-    public function getSprintReportProjects(): SprintReportProjects
-    {
-        $sprintReportProjects = new SprintReportProjects();
-        $projects = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.projects.getAll', []);
-
-        foreach ($projects as $project) {
-            $sprintReportProject = new SprintReportProject();
-            $sprintReportProject->id = $project->id;
-            $sprintReportProject->name = $project->name;
-
-            $sprintReportProjects->projects->add($sprintReportProject);
-        }
-
-        return $sprintReportProjects;
-    }
-
-    /**
-     * Get projectV2.
-     *
-     * @param string $projectId A project key or id
-     *
-     * @return SprintReportProject SprintReportProject
-     *
-     * @throws ApiServiceException|EconomicsException
-     */
-    public function getSprintReportProject(string $projectId): SprintReportProject
-    {
-        $project = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.projects.getProject', ['id' => $projectId]);
-
-        $sprintReportProject = new SprintReportProject();
-        $sprintReportProject->id = $project->id;
-        $sprintReportProject->name = $project->name;
-
-        return $sprintReportProject;
-    }
-
-    /**
      * @throws ApiServiceException
-     * @throws EconomicsException
      */
     private function getProjectIssuesPaged($projectId, $startAt, $maxResults = 50): array
     {
@@ -163,7 +104,6 @@ class LeantimeApiService implements DataProviderServiceInterface
 
     /**
      * @throws ApiServiceException
-     * @throws EconomicsException
      */
     public function getIssuesDataForProjectPaged(string $projectId, $startAt = 0, $maxResults = 50): PagedResult
     {
@@ -246,7 +186,7 @@ class LeantimeApiService implements DataProviderServiceInterface
             $projectData->projectTrackerKey = $project->id;
             $projectData->projectTrackerProjectUrl = $this->leantimeUrl.'#/tickets/showTicket/'.$project->id;
 
-            $projectVersions = $this->getSprintReportVersions($project->id);
+            $projectVersions = $this->getProjectVersions($project->id);
             foreach ($projectVersions as $projectVersion) {
                 $projectData->versions?->add($projectVersion);
             }
@@ -280,34 +220,21 @@ class LeantimeApiService implements DataProviderServiceInterface
         return [];
     }
 
-    /**
-     * Get milestone.
-     *
-     * @param $key
-     *   A milestone key or id
-     *
-     * @throws ApiServiceException
-     */
-    public function getMilestone($key): mixed
+    public function getProjectVersions(string $projectId): Versions
     {
-        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getTicket', ['id' => $key]);
-    }
-
-    public function getSprintReportVersions(string $projectId): SprintReportVersions
-    {
-        $sprintReportVersions = new SprintReportVersions();
+        $versions = new Versions();
         $projectVersions = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAllMilestones', ['searchCriteria' => ['currentProject' => $projectId, 'type' => 'milestone']]);
 
         foreach ($projectVersions as $projectVersion) {
-            $sprintReportVersion = new SprintReportVersion();
-            $sprintReportVersion->id = $projectVersion->id;
-            $sprintReportVersion->name = $projectVersion->headline;
-            $sprintReportVersion->projectTrackerId = $projectVersion->projectId;
+            $version = new VersionModel();
+            $version->id = $projectVersion->id;
+            $version->name = $projectVersion->headline;
+            $version->projectTrackerId = $projectVersion->projectId;
 
-            $sprintReportVersions->versions->add($sprintReportVersion);
+            $versions->versions->add($version);
         }
 
-        return $sprintReportVersions;
+        return $versions;
     }
 
     public function getWorklogDataCollection(string $projectId): WorklogDataCollection
@@ -344,70 +271,6 @@ class LeantimeApiService implements DataProviderServiceInterface
         }
 
         return $worklogDataCollection;
-    }
-
-    /**
-     * Get all sprints for a given board.
-     *
-     * @throws ApiServiceException
-     */
-    public function getAllSprints(): array
-    {
-        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.sprints.getAllSprints', ['projectId' => '6']);
-    }
-
-    /**
-     * Get all issues for given board and sprint.
-     *
-     * @param string $sprintId id of the sprint to extract issues for
-     *
-     * @return array array of issues
-     *
-     * @throws ApiServiceException
-     */
-    public function getTicketsInSprint(string $sprintId): array
-    {
-        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['sprint' => $sprintId]]);
-    }
-
-    /**
-     * @throws ApiServiceException
-     */
-    private function getIssueSprint($issueEntry): SprintReportSprint
-    {
-        $sprint = false;
-
-        if ((bool) $issueEntry->sprint) {
-            $sprint = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.sprints.getSprint', ['id' => $issueEntry->sprint]);
-        }
-
-        if ($sprint) {
-            $sprintState = SprintStateEnum::OTHER;
-            $sprintTemporal = $this->getDateSpanTemporal($sprint->startDate, $sprint->endDate);
-
-            switch ($sprintTemporal) {
-                case self::PAST:
-                    $sprintState = SprintStateEnum::OTHER;
-                    break;
-                case self::PRESENT:
-                    $sprintState = SprintStateEnum::ACTIVE;
-                    break;
-                case self::FUTURE:
-                    $sprintState = SprintStateEnum::FUTURE;
-                    break;
-            }
-
-            return new SprintReportSprint(
-                $sprint->id,
-                $sprint->name,
-                $sprintState,
-                $sprint->startDate ? strtotime($sprint->startDate) : null,
-                $sprint->endDate ? strtotime($sprint->endDate) : null,
-                null,
-            );
-        } else {
-            throw new ApiServiceException('Sprint not found', 404);
-        }
     }
 
     /**
@@ -632,13 +495,6 @@ class LeantimeApiService implements DataProviderServiceInterface
         return $planning;
     }
 
-    public function getTimesheetsForTicket(string $ticketId): mixed
-    {
-        return $this->getTimesheets([
-            'ticketFilter' => $ticketId,
-        ]);
-    }
-
     private function getTimesheets(array $params): mixed
     {
         return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.timesheets.getAll', $params + [
@@ -653,156 +509,9 @@ class LeantimeApiService implements DataProviderServiceInterface
     }
 
     /**
-     * @throws ApiServiceException
-     */
-    private function getIssuesForProjectMilestone($projectId, $milestoneId): array
-    {
-        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['currentProject' => $projectId, 'milestone' => $milestoneId]]);
-    }
-
-    /**
-     * @throws ApiServiceException
-     * @throws \Exception
-     */
-    public function getSprintReportData(string $projectId, string $versionId): SprintReportData
-    {
-        $sprintReportData = new SprintReportData();
-        $epics = $sprintReportData->epics;
-        $issues = $sprintReportData->issues;
-        $sprints = $sprintReportData->sprints;
-
-        $spentSum = 0;
-        $remainingSum = 0;
-
-        $epics->set('noEpic', new SprintReportEpic('noEpic', 'Uden Tag'));
-
-        // Get version and project.
-        $milestone = $this->getMilestone($versionId);
-        $project = $this->getProject($projectId);
-
-        $issueEntries = $this->getIssuesForProjectMilestone($projectId, $versionId);
-
-        $issueCount = 1;
-        foreach ($issueEntries as $issueEntry) {
-            $issue = new SprintReportIssue();
-            $issues->add($issue);
-
-            /* Tags are stored as a comma seperated string.
-            In our implementation of Leantime, tickets are only supposed to have one tag.
-            Tickets with multiple tags will not break, but it would look wierd in the report. */
-            if (isset($issueEntry->tags)) {
-                $tag = new SprintReportEpic($issueEntry->tags, $issueEntry->tags);
-                $epics->set($issueEntry->tags, $tag);
-            } else {
-                $tag = $epics->get('noEpic');
-            }
-
-            if (!$tag instanceof SprintReportEpic) {
-                continue;
-            }
-
-            $issue->epic = $tag;
-
-            // Get sprint for issue.
-            if ((bool) $issueEntry->sprint) {
-                try {
-                    $issueSprint = $this->getIssueSprint($issueEntry);
-
-                    if (!$sprints->containsKey($issueSprint->id)) {
-                        $sprints->set($issueSprint->id, $issueSprint);
-                    }
-                    // Set which sprint the issue is assigned to.
-                    if (SprintStateEnum::ACTIVE === $issueSprint->state || SprintStateEnum::FUTURE === $issueSprint->state) {
-                        $issue->assignedToSprint = $issueSprint;
-                    }
-                } catch (ApiServiceException) {
-                    // Ignore if sprint is not found.
-                }
-            }
-            $worklogs = $this->getTimesheetsForTicket($issueEntry->id);
-
-            foreach ($worklogs as $worklog) {
-                $workLogStarted = strtotime($worklog->workDate);
-
-                $worklogSprints = array_filter($sprints->toArray(), function ($sprintEntry) use ($workLogStarted) {
-                    /* @var SprintReportSprint $sprintEntry */
-                    return
-                        $sprintEntry->startDateTimestamp <= $workLogStarted
-                        && ($sprintEntry->completedDateTimstamp ?? $sprintEntry->endDateTimestamp) > $workLogStarted;
-                });
-
-                $worklogSprintId = self::NO_SPRINT;
-
-                if (!empty($worklogSprints)) {
-                    $worklogSprint = array_pop($worklogSprints);
-                    $worklogSprintId = $worklogSprint->id;
-                }
-                $newLoggedWork = (float) ($issue->epic->loggedWork->containsKey($worklogSprintId) ? $issue->epic->loggedWork->get($worklogSprintId) : 0) + ($worklog->hours * 60 * 60);
-                $issue->epic->loggedWork->set($worklogSprintId, $newLoggedWork);
-            }
-
-            // Accumulate spentSum.
-            $spentSum += ($issueEntry->bookedHours * 60 * 60);
-            $issue->epic->spentSum += ($issueEntry->bookedHours * 60 * 60);
-
-            // Accumulate remainingSum.
-            if ('0' !== $issueEntry->status && isset($issueEntry->hourRemaining)) {
-                $remainingEstimateSeconds = ($issueEntry->hourRemaining * 60 * 60);
-                $remainingSum += $remainingEstimateSeconds;
-
-                $issue->epic->remainingSum += $remainingEstimateSeconds;
-
-                $assignedToSprint = $issue->assignedToSprint;
-                $newRemainingWork = (float) ($issue->epic->remainingWork->containsKey($assignedToSprint->id) ? $issue->epic->remainingWork->get($assignedToSprint->id) : 0) + $remainingEstimateSeconds;
-                $issue->epic->remainingWork->set($assignedToSprint->id, $newRemainingWork);
-                $issue->epic->plannedWorkSum += $remainingEstimateSeconds;
-            }
-
-            // Accumulate originalEstimateSum.
-            if (isset($issueEntry->planHours)) {
-                $issue->epic->originalEstimateSum += ($issueEntry->planHours * 60 * 60);
-
-                $sprintReportData->originalEstimateSum += ($issueEntry->planHours * 60 * 60);
-            }
-            ++$issueCount;
-        }
-
-        // Sort sprints by key.
-        /** @var \ArrayIterator $iterator */
-        $iterator = $sprints->getIterator();
-        $iterator->uasort(function ($a, $b) {
-            return mb_strtolower($a->id) <=> mb_strtolower($b->id);
-        });
-        $sprints = new ArrayCollection(iterator_to_array($iterator));
-
-        // Sort epics by name.
-        /** @var \ArrayIterator $iterator */
-        $iterator = $epics->getIterator();
-        $iterator->uasort(function ($a, $b) {
-            return mb_strtolower($a->name) <=> mb_strtolower($b->name);
-        });
-        $epics = new ArrayCollection(iterator_to_array($iterator));
-        // Calculate spent, remaining hours.
-        $spentHours = $spentSum / 3600;
-        $remainingHours = $remainingSum / 3600;
-
-        $sprintReportData->projectName = $project->name;
-        $sprintReportData->versionName = $milestone->headline;
-        $sprintReportData->remainingHours = $remainingHours;
-        $sprintReportData->spentHours = $spentHours;
-        $sprintReportData->spentSum = $spentSum;
-        $sprintReportData->projectHours = $spentHours + $remainingHours;
-        $sprintReportData->epics = $epics;
-        $sprintReportData->sprints = $sprints;
-
-        return $sprintReportData;
-    }
-
-    /**
      * Get from Leantime.
      *
      * @throws ApiServiceException
-     * @throws EconomicsException
      */
     private function request(string $path, string $type, string $method, array $params = []): mixed
     {
@@ -836,20 +545,6 @@ class LeantimeApiService implements DataProviderServiceInterface
         }
 
         return null;
-    }
-
-    private function getDateSpanTemporal($startDate, $endDate): string
-    {
-        $currentDate = time();
-        $startDate = strtotime($startDate);
-        $endDate = strtotime($endDate);
-        if ($startDate < $currentDate && $endDate > $currentDate) {
-            return self::PRESENT;
-        } elseif ($startDate < $currentDate && $endDate < $currentDate) {
-            return self::PAST;
-        } else {
-            return self::FUTURE;
-        }
     }
 
     private function getLeantimeTimeZone(): \DateTimeZone